LIBWEB-2456. Added a Tomcat environment for Hippo.
Separated the Hippo deployment out into a separate script, hippo.sh, that runs unprivileged (as the vagrant user). In addition to installing the diet and server-diet tarballs from the Hippo source, it also makes sure that the PostgreSQL JDBC JAR is installed in the common/lib directory (if necessary, it downloads it from the Postgres website).

Included a Tomcat environment copied and slightly modified from wwwdev. Modifications include turning off proxying for the 9600 and 9603-9605 connectors.

Updated the README with instructions on building Hippo and deploying it to the Vagrant.

https://issues.umd.edu/browse/LIBWEB-2456