LIBWEB-2245. Split Hippo between 2 Tomcats.
Created parallel structures for the site and cms. Increased the VM total memory to 4GB, and reduced the max heap size for each Tomcat JVM to 1GB. Also separated out the provisioning of the tomcat runtime environments into a separate script, tomcat.sh.

Bumped the expected Hippo version up to 7.8.8-4-SNAPSHOT, to match the develop branch on umd-lib/public-hippo.

https://issues.umd.edu/browse/LIBWEB-2245