LIBWEB-2495. Add cluster configuration to the common repository.xml.
Add the Cluster configuration element to the common Tomcat environment config files. This will get copied to both the tomcat-site and tomcat-cms runtime directories by the tomcat.sh provisioning script. Note that we do not need to explicitly set a cluster ID, as it appears that Jackrabbit automatically generates it as a UUID and places it in the {$STORAGE_DIR}/cluster_node.id file.

https://issues.umd.edu/browse/LIBWEB-2495