LIBWEB-2245. Add site-Tomcat-specific options to startup.
Moved the definition of the Hippo, Log4J, and Catalina options variables into the env-config.properties files for site and cms. Then added the "-Drepo.bootstrap=false -Dhippo.scheduler.disabled=true" flags to the site.

Note too that the command to load the env-config.properties is now written out as "source conf/env-config.properties" instead of just using the Bash "." command. This should hopefully improve readability of the script.

https://issues.umd.edu/browse/LIBWEB-2245